Move errors to http package
