+++ internal/clients/http/common_http_errors.go
