+++ internal/clients/http/ec2_errors.go
@@ -1,0 +1,5 @@
+package http
+
+import "errors"
+
+var DuplicatePubkeyErr = errors.New("pubkey already exists in the target account")
