+++ internal/clients/http/image_builder_errors.go
@@ -1,4 +1,4 @@
-package image_builder
+package http
 
 import (
 	"errors"
