+++ internal/clients/http/sources_errors.go
@@ -1,4 +1,4 @@
-package sources
+package http
 
 import (
 	"errors"
